@@ -24,12 +24,14 @@ import android.annotation.TargetApi;
 import android.content.Intent;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.text.ClipboardManager;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 @TargetApi(Build.VERSION_CODES.JELLY_BEAN)
 @SuppressLint("NewApi")
@@ -43,7 +45,7 @@ public class ZANaviDonateActivity extends AppCompatActivity
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.activity_donate);
 
-		android.support.v7.widget.Toolbar bar = (android.support.v7.widget.Toolbar) findViewById(R.id.toolbar2bd);
+		Toolbar bar = (Toolbar) findViewById(R.id.toolbar2bd);
 		bar.setTitle(Navit.get_text("Donate with Bitcoin"));
 		bar.setNavigationOnClickListener(new View.OnClickListener()
 		{
