@@ -27,11 +27,13 @@ import android.content.pm.PackageManager.NameNotFoundException;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 @TargetApi(Build.VERSION_CODES.JELLY_BEAN)
 @SuppressLint("NewApi")
@@ -49,7 +51,7 @@ public class ZANaviNormalDonateActivity extends AppCompatActivity
 
 		setContentView(R.layout.activity_normal_donate);
 
-		android.support.v7.widget.Toolbar bar = (android.support.v7.widget.Toolbar) findViewById(R.id.toolbar2nd);
+		Toolbar bar = (Toolbar) findViewById(R.id.toolbar2nd);
 		bar.setTitle(Navit.get_text("Donate"));
 		bar.setNavigationOnClickListener(new View.OnClickListener()
 		{
